sql: refactor render function for `ShowCreateTrigger()`
Moved `renderCreateTriggerStatement()` into
`show_trigger.go` and used it in `ShowCreateTrigger()`
to avoid repeating the code for generating trigger
create statements.
Also, removes the work-around explicitly
iterating over query, as the SQL query execution
bug is fixed.

Release note: None
Epic: CRDB-49582
