+++ pkg/sql/conn_executor_exec.go
@@ -3458,8 +3458,7 @@ func (ex *connExecutor) makeExecPlan(
 			// session var is set to 'true' ('false' is the default).
 			catalog = planner.optPlanningCtx.catalog
 		}
-		planStr := ih.planGist.String()
-		_, err := explain.DecodePlanGistToRows(ctx, &planner.extendedEvalCtx.Context, planStr, catalog)
+		_, err := explain.DecodePlanGistToRows(ctx, &planner.extendedEvalCtx.Context, ih.planGist.String(), catalog)
 		if err != nil {
 			return ctx, errors.NewAssertionErrorWithWrappedErrf(err, "failed to decode plan gist: %q", ih.planGist.String())
 		}
