+++ pkg/sql/crdb_internal.go
@@ -3944,91 +3944,6 @@ CREATE TABLE crdb_internal.create_procedure_statements (
 	populate: createRoutinePopulate(true /* procedure */),
 }
 
-func renderCreateTriggerStatement(
-	ctx context.Context,
-	p *planner,
-	trigger *descpb.TriggerDescriptor,
-	tableDesc catalog.TableDescriptor,
-) (string, error) {
-	tableTyp, err := p.ResolveTypeByOID(ctx, typedesc.TableIDToImplicitTypeOID(tableDesc.GetID()))
-	if err != nil {
-		return "", err
-	}
-
-	funcName, err := p.GetQualifiedFunctionNameByID(ctx, int64(trigger.FuncID))
-	if err != nil {
-		return "", err
-	}
-
-	tableName, err := p.getQualifiedTableName(ctx, tableDesc)
-	if err != nil {
-		return "", err
-	}
-
-	events := make([]*tree.TriggerEvent, len(trigger.Events))
-	for j := range events {
-		descEvent := trigger.Events[j]
-		events[j] = &tree.TriggerEvent{
-			EventType: tree.TriggerEventTypeToTree[descEvent.Type],
-			Columns:   make(tree.NameList, 0, len(descEvent.ColumnNames)),
-		}
-		for _, colName := range descEvent.ColumnNames {
-			events[j].Columns = append(events[j].Columns, tree.Name(colName))
-		}
-	}
-
-	var transitions []*tree.TriggerTransition
-	if trigger.NewTransitionAlias != "" {
-		transitions = append(transitions, &tree.TriggerTransition{
-			IsNew: true,
-			Name:  tree.Name(trigger.NewTransitionAlias),
-		})
-	}
-	if trigger.OldTransitionAlias != "" {
-		transitions = append(transitions, &tree.TriggerTransition{
-			IsNew: false,
-			Name:  tree.Name(trigger.OldTransitionAlias),
-		})
-	}
-
-	forEach := tree.TriggerForEachStatement
-	if trigger.ForEachRow {
-		forEach = tree.TriggerForEachRow
-	}
-
-	var whenExpr tree.Expr
-	if trigger.WhenExpr != "" {
-		whenExpr, err = schemaexpr.ParseTriggerWhenExprForDisplay(
-			ctx, tableTyp, trigger.WhenExpr, p.EvalContext(), p.SemaCtx(), tree.FmtParsable,
-		)
-		if err != nil {
-			return "", err
-		}
-	}
-
-	createTrigger := &tree.CreateTrigger{
-		Replace:     false,
-		Name:        tree.Name(trigger.Name),
-		ActionTime:  tree.TriggerActionTimeToTree[trigger.ActionTime],
-		Events:      events,
-		TableName:   tableName.ToUnresolvedObjectName(),
-		Transitions: transitions,
-		ForEach:     forEach,
-		When:        whenExpr,
-		FuncName:    funcName.ToUnresolvedObjectName().ToUnresolvedName(),
-		FuncArgs:    trigger.FuncArgs,
-	}
-
-	f := tree.NewFmtCtx(
-		tree.FmtParsable,
-		tree.FmtDataConversionConfig(p.SessionData().DataConversionConfig),
-		tree.FmtLocation(p.SessionData().Location),
-	)
-	f.FormatNode(createTrigger)
-
-	return f.CloseAndGetString(), nil
-}
-
 func createTriggerPopulate(
 	ctx context.Context, p *planner, db catalog.DatabaseDescriptor, addRow func(...tree.Datum) error,
 ) error {
